[td] Major refactor of datasets slug page
